PSS-41 | Поправил компоненты Clock/Score / Правки от ревьюера
