+++ packages/client/src/shared/components/clock/styles.module.css
@@ -10,8 +10,22 @@
   align-items: center;
 }
 
-.score {
+.clock {
   font-family: 'Roboto', sans-serif;
   font-size: 18px;
   color: #fff;
+}
+
+.separator {
+  animation: blink 2s step-end infinite;
+}
+
+@keyframes blink {
+  0%,
+  100% {
+    opacity: 1;
+  }
+  50% {
+    opacity: 0;
+  }
 }
