+++ packages/client/src/shared/constants/icons.ts
@@ -1,0 +1,19 @@
+import Clock from '@/assets/images/icons/clock.svg'
+import Gil from '@/assets/images/icons/gil.svg'
+import Magic from '@/assets/images/icons/magic.svg'
+import MagicDefense from '@/assets/images/icons/magicDefense.svg'
+import Necklace from '@/assets/images/icons/necklace.svg'
+import Spirit from '@/assets/images/icons/spirit.svg'
+import Strength from '@/assets/images/icons/strength.svg'
+import Vitality from '@/assets/images/icons/vitality.svg'
+
+export const ICONS = {
+  Clock,
+  Gil,
+  Magic,
+  MagicDefense,
+  Necklace,
+  Spirit,
+  Strength,
+  Vitality,
+}
